replace "as.integer( as.character( ggData$year ) )" by "as.numeric( as.character( ggData$year ) )", because in some cases further calculations with this integer variable (e.g. exp()) result in integer variables although they are actually not integer variables
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/micecon@2050 95fa4d17-b1e8-0310-9d42-a43d67248e42

+++ pkg/micEcon/tests/translog.R
@@ -463,7 +463,7 @@ ggElaRanLogMet <- elas( ggResultRanLog )
 all.equal( ggElaRanMet, ggElaRanLogMet, check.attributes = FALSE )
 
 ## panel data with a shifter
-ggData$yearInt <- as.integer( as.character( ggData$year ) )
+ggData$yearInt <- as.numeric( as.character( ggData$year ) )
 ggData$tech <- exp( ggData$yearInt - min( ggData$yearInt ) )
 # fixed effects
 ggResShifter <- translogEst( "invest", c( "value", "capital" ), ggData,
